- Actually set the key/limit to 0 on -k / -l as join requires this, turning   off the mode flag is not enough. - Adding some of the code which will eventually be used to allow a test-suite   module.
git-svn-id: http://svn.ratbox.org/svnroot/trunk@7789 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/client.c
@@ -68,6 +68,9 @@ static EVH check_pings;
 static int remote_client_count=0;
 static int local_client_count=0;
 
+void (*module_connect_close)(struct Client *who);
+// void (*module_
+
 /*
  * init_client
  *
@@ -1292,10 +1295,12 @@ const char* comment         /* Reason for the exit */
             free_dlink_node(m);
            }
         }
+  if (!IsModuleOwned(source_p))
+  {
       delete_adns_queries(source_p->localClient->dns_query);
       delete_identd_queries(source_p);
-
       client_flush_input(source_p);
+  }
 
       /* This source_p could have status of one of STAT_UNKNOWN, STAT_CONNECTING
        * STAT_HANDSHAKE or STAT_UNKNOWN
@@ -1393,8 +1398,10 @@ const char* comment         /* Reason for the exit */
       ** It also makes source_p->from == NULL, thus it's unnecessary
       ** to test whether "source_p != target_p" in the following loops.
       */
-      
-      close_connection(source_p);
+      if (!IsModuleOwned(source_p))
+        close_connection(source_p);
+      else if (module_connect_close)
+        module_connect_close(source_p);
     }
 
   if(IsServer(source_p))
